annotate missing ui requirements
