+++ packages/nextjs/app/lobby/page.tsx
@@ -1,0 +1,4 @@
+export default function LobbyPage() {
+  // TODO: implement lobby selection view (Action Plan 1.1)
+  return <div>Lobby TODO</div>;
+}
